Added cinematicShooting.sqf
This script creates a "cinematic shooting" scenario between all units that are synchronized to the parameter object, meaning they will be able to fire more or less non-stop at each other without killing anyone or running out of ammunition. The script ends when the synchObject is deleted. Recommend synchObject is a game logic object.